+++ ci/build_linux_conda.sh
@@ -1,6 +1,11 @@
 if [ -n "$TRAVIS_TAG_DSS_PYTHON" ]; then # only run conda-build on tags, takes too long
     # Install Miniconda
-    wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh -q
+    if [ -n "$LINUX_X64" ]; then
+        wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh -q
+    else
+        wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86.sh -O miniconda.sh -q
+    fi
+    
     bash miniconda.sh -b -p $HOME/miniconda
     export PATH="$HOME/miniconda/bin:$PATH"
     conda config --set always_yes yes --set changeps1 no