+++ ci/upload_windows.sh
@@ -4,3 +4,6 @@ if [ "$DSS_PYTHON_BUILD_TAG" == "1" ]; then
         find ../artifacts -name "*.whl" -or -name "*.tar.bz2" | xargs -I {} anaconda upload --no-progress -l main -u pmeira {}
     fi
 fi
+
+# Always upload to AppVeyor
+find ../artifacts -name "*.whl" -or -name "*.tar.bz2" | xargs -I {} appveyor PushArtifact {}